+++ docs/releases.rst
@@ -5,6 +5,7 @@ Version 3.4, To Be Released
 ---------------------------
 
 * Deprecate ``--tap-stream`` in favor of ``--tap`` for streaming mode.
+* When using xdist, report the plan at the beginning of execution.
 * Add support for Python 3.11.
 * Drop support for Python 3.7 (it is end-of-life).
 * Drop support for Python 3.6 (it is end-of-life).